requirements.txt
# Medical Image Viewer - MedVision
# Requirements for Python 3.8+

streamlit==1.32.0
pydicom==2.4.4
numpy==1.26.2
pillow==10.2.0
opencv-python==4.8.1.78
pandas==2.1.3
sqlalchemy==2.0.25